Revert "Merge branch 'main' into HUD"
This reverts commit e61148d4a5b9a7ff84f2fe984b0ac06e2e866c74, reversing
changes made to 4ffbfd2be6a36819b8b306c5109d3f2cd77b9854.

+++ Assets/Code/Entity/Enemy Scripts/Enemy.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 // this is just so that we have a base class for stats and save on rewriting code
 public class Enemy : MonoBehaviour//, IOnBulletHit
 {
-    public int health = 100;
+    public double health = 100;
     public int armor = 0;
 
     // Start is called once before the first execution of Update after the MonoBehaviour is created
